ingestion microservice polish

+++ Domain/Models/TweetDto.cs
@@ -5,8 +5,8 @@ namespace Domain.Models
 {
     public class TweetDto
     {
-        public int TwitterId { get; set; }
-        public int AuthorId { get; set; }
+        public long TwitterId { get; set; }
+        public long AuthorId { get; set; }
         public string Text { get; set; }
         public DateTime CreatedAt { get; set; }
         public List<HashtagDto> Hashtags { get; set; }
