+++ TwitterFeed/Contracts/TweetVm.cs
@@ -5,8 +5,8 @@ namespace TwitterFeed.Contracts
 {
     public class TweetVm
     {
-        public int TwitterId { get; set; }
-        public int AuthorId { get; set; }
+        public long TwitterId { get; set; }
+        public long AuthorId { get; set; }
         public string Text { get; set; }
         public DateTime CreatedAt { get; set; }
         public List<HashtagVm> Hashtags { get; set; }
